[URP] Upscaling Filters & FXAA Scaling Logic
* Move FXAA Logic to Common.hlsl

This change moves the FXAA HLSL implementation into the common post
processing shader file. This will make it easier to execute the FXAA
shader logic outside of the FinalPost pass which is a prerequisite for
FSR.

This change also modifies the FXAA Load helper function to make it use
point sampling instead of linear sampling on GLES. This should yield
more consistent behavior between GLES and non-GLES environments.

* Upscaling Filter Controls

This change adds a new property to the pipeline asset that allows users
to control which filter is used when upscaling is performed. The current
implementation supports selecting either bilinear or nearest-neighbor.
Additional methods of filtering will be added in future changes.

* Converted isScaledRender to an Enum

This change adds a new enum called ImageScaling which enumerates all
possible image scaling scenarios. This helps make the scaling related
conditional logic easier to read.

* Add Automatic Upscaling Filter

This change splits the upscaling filter selected by the user from the
one used within URP's implementation to allow for a "meta filter"
called Auto. This new filter automatically switches between bilinear
and nearest-neighbor filtering based on the current rendering
environment's ability to perform integer scaling.

* Rename Point Sampling Shader Macro

This change renames the _FILTER_POINT macro to _POINT_SAMPLING to
improve consistency with future FSR changes.

* Fix XR Compatibility for FXAA Common Code

This change updates the FXAA common shader functions to support 2d array
input textures which are used in XR's single pass instanced mode.

* Upscaling Filter Tests for URP

This commit adds new test cases for the upscaling filters that were
added in a previous change.

* Replace TemporaryRT with RTHandle

This change updates the upscaling setup logic to use an RTHandle instead
of a temporary render target.

* Updated Changelog

Updated the URP changelog to reflect the upscaling filter changes and
the FXAA scaling fix.

* Fix Black Screen for XR

This change resolves a black screen issue on XR platforms
(tested in the mock HMD) which was caused by missing support for
draw procedural in the upscaling setup shader.

* Revert Auto-Format Changes

This commit reverts the auto-format changes made in
PostProcessData.asset.

* Upscale Setup Shader Naming Consistency Fix

This commit renames upscaleSetupPs -> upscaleSetupPS to make the
naming consistent with other shaders.

* Apply Review Feedback

This commit addresses various pieces of review feedback such as missing
docs, renames, and data visibility. It also fixes a draw procedural
related multi-compile option.

* Comment Updates for UpscaleSetup.shader

This change adds a comment block in the upscaling setup shader that
describes the cases where it's used and why.

* Renamed UpscaleSetup to ScalingSetup

This change renames the extra scaling blit pass to better communicate
that it can be used in both upscaling and downscaling scenarios.

* Reworded Changelog Notes

This change adds additional information to the changelog notes for the
upscaling filters.