Fix Interaction with Material Overrides
This change updates the Alpha-To-Coverage logic to handle the case where
material override functionality is enabled. In this case, we aren't able
to set the AlphaToMask render state based off of the shader graph target
alone since it's possible for it to be changed later via overrides. To
solve this, we fall back to using the hidden _AlphaToMask material
property similar to the non-graph shaders.