Fix banding in MultiScaleVolumetricObscurance (AmbientOcclusion)
The precision of the temporary TileDepth* rendertargets are lower than all other temporary depth target when applying MSVO; With some specific camera angle/distances, this can cause banding/lines effects to appear due to these low precision buffer.

This change will add a user-selectable option to allow to increase the precision of the TileDepth* buffers. This solves the banding issue at the cost of GPU memory and GPU bandwidth.