Sorting Network Improvements & Debug Visualization
This change converts the existing 4 input sorting network into an 8
input sorting network so it can handle more of the common cases. This
change also cleans up the network code a bit but it's now using indexed
temps based on the dxbc which need to be removed if possible since they
can cause scratch memory usage in the compiler.

A sorting debug visualization was also added as part of this change
which shows incorrectly sorted samples as red, correctly sorted samples
as green, and samples that could not be sorted correctly due to depth
precision as yellow.