Switch to Sorted Sample Indices
This change modifies the existing sorting implementation to produce the
sorted sample indices as a 32-bit index list instead of perform sample
data re-ordering as part of the sort shader. This improves the
performance of the sort since it no longer needs to move large sample
data around during sorting. It does have a slight negative effect on the
sample accumulation performance due to the extra indirection. Overall,
the change results in a net win for the performance of a frame.

In the future, the index list should be updated to use 8 bits per sorted
index instead of 32.
(Assuming we never need more than 256 samples per pixel)