Fix Sample Sorting in Debug View
This change updates the fullscreen debug view code to use sorted sample
index values instead of just walking unordered sample lists.

Note: A read-only binding for sort memory was added the common OIT
bindings but only in debug display shaders. This allows us to bind it as
a normal buffer instead of needing to use a UAV. (Debug display is
always performed as a draw operation rather than a dispatch)