Misc Sorting Improvements
This change updates the network sorts to remove all dynamic indexing
into temporary arrays which allows the FXC compiler to lift all indexed
temps into registers.

This commit also updates the sort keys to ensure that valid keys are
always sorted towards the beginning so we can perform contiguous writes
across lanes when writing the sorted indices at the end of the shader.