Update TextureGather Macros in GLCore.hlsl (#7029)
* Update TextureGather Macros in GLCore.hlsl

This change updates the platform shader macros in GLCore.hlsl to enable
support for TextureGather operations on Unity's 4.5 shader target.

The changes in this PR are based off of GH #2346 which modified the
GLES3.hlsl file in order to fix cubemap array support. This PR basically
takes those same changes and applies them to GLCore.hlsl in addition.

This commit also adds some missing component specific macros for
texture gather (red, green, blue, alpha).

* Update Changelog

This commit updates the changelog to reflect the recent changes to the
GLCore.hlsl header.