Address Review Feedback
This change restores the AlphaDiscard() function and replaces its
internals with the new alpha-to-coverage logic. This makes several small
shader changes no longer necessary and makes the diff a bit smaller (in
addition to not breaking existing external code). This change also
attempts to rename the AlphaToMaskEnabled shader variables to better
reflect their true meaning.