Fix Regression in GLCore from GH #7029 (#7062)
* Fix Regression in GLCore from GH #7029

The changes in GH #7029 accidentally broke the platform shader macro
override system for GLCore's texture cube arrays. This change fixes
the issue by restoring the missing platform macros.

* Fall Back to UNITY_NO_CUBEMAP_ARRAY Due to Bug

Due to some sort of bug, SHADER_AVAILABLE_CUBEARRAY is not defined in
compute shaders. This commit switches the code back to
UNITY_NO_CUBEMAP_ARRAY to avoid futher problems.

(FogBugz Case 1402457)

* Fix Comment Typo

* Revert UNITY_NO_CUBEMAP_ARRAY Changes

This change removes usage of UNITY_NO_CUBEMAP_ARRAY and replaces it with
SHADER_AVAILABLE_CUBEARRAY since that's the correct macro to be using.
The bug we encountered after this change in VFX was actually due to a
missing pragma: "#pragma require cubearray".