More Sorting Optimizations
Added Sorting Network paths for <= 16 and <= 32 sample counts. It turns
out that this is signficiantly faster than the wave based radix sort for
32 elements which wasn't technically safe anyways. This cuts sorting
time from ~1.6ms to ~600us in extreme cases.

This change also removes the wave radix sort path and cleans up the 256
item radix sort path a bit (and makes it use 8 bits).
It still has a serial section though. :(
That should be fixed eventually.