Backport URP FSR to 2021.2 (#6939)
* Common FSR Utilities in SRP Core

* Common FSR Utilities for SRPs

This change adds two new utility files to SRP core to make it easier
to use FSR in SRPs.

The FSRUtils C# class contains a port of the FSR constant values setup
code from the HLSL files. This is needed for SRPs that do not support
writing to buffers from inside a shader.

The FSRCommon HLSL include contains helper functions for using EASU
and RCAS inside the Unity shader environment.

The two files must be used together since they are tightly coupled by
the identifiers used for the shader constants.

* [URP] Upscaling Filters & FXAA Scaling Logic

* Move FXAA Logic to Common.hlsl

This change moves the FXAA HLSL implementation into the common post
processing shader file. This will make it easier to execute the FXAA
shader logic outside of the FinalPost pass which is a prerequisite for
FSR.

This change also modifies the FXAA Load helper function to make it use
point sampling instead of linear sampling on GLES. This should yield
more consistent behavior between GLES and non-GLES environments.

* Upscaling Filter Controls

This change adds a new property to the pipeline asset that allows users
to control which filter is used when upscaling is performed. The current
implementation supports selecting either bilinear or nearest-neighbor.
Additional methods of filtering will be added in future changes.

* Converted isScaledRender to an Enum

This change adds a new enum called ImageScaling which enumerates all
possible image scaling scenarios. This helps make the scaling related
conditional logic easier to read.

* Add Automatic Upscaling Filter

This change splits the upscaling filter selected by the user from the
one used within URP's implementation to allow for a "meta filter"
called Auto. This new filter automatically switches between bilinear
and nearest-neighbor filtering based on the current rendering
environment's ability to perform integer scaling.

* Rename Point Sampling Shader Macro

This change renames the _FILTER_POINT macro to _POINT_SAMPLING to
improve consistency with future FSR changes.

* Fix XR Compatibility for FXAA Common Code

This change updates the FXAA common shader functions to support 2d array
input textures which are used in XR's single pass instanced mode.

* Upscaling Filter Tests for URP

This commit adds new test cases for the upscaling filters that were
added in a previous change.

* Replace TemporaryRT with RTHandle

This change updates the upscaling setup logic to use an RTHandle instead
of a temporary render target.

* Updated Changelog

Updated the URP changelog to reflect the upscaling filter changes and
the FXAA scaling fix.

* Fix Black Screen for XR

This change resolves a black screen issue on XR platforms
(tested in the mock HMD) which was caused by missing support for
draw procedural in the upscaling setup shader.

* Revert Auto-Format Changes

This commit reverts the auto-format changes made in
PostProcessData.asset.

* Upscale Setup Shader Naming Consistency Fix

This commit renames upscaleSetupPs -> upscaleSetupPS to make the
naming consistent with other shaders.

* Apply Review Feedback

This commit addresses various pieces of review feedback such as missing
docs, renames, and data visibility. It also fixes a draw procedural
related multi-compile option.

* Comment Updates for UpscaleSetup.shader

This change adds a comment block in the upscaling setup shader that
describes the cases where it's used and why.

* Renamed UpscaleSetup to ScalingSetup

This change renames the extra scaling blit pass to better communicate
that it can be used in both upscaling and downscaling scenarios.

* Reworded Changelog Notes

This change adds additional information to the changelog notes for the
upscaling filters.

* [URP] Port HDRP Mip Bias Logic to URP

* Port HDRP Mip Bias Logic to URP

This change integrates the mip bias logic from HDRP into URP. This logic
ensures that shaders select their mips based on the final screen
resolution rather than the current render resolution. In cases where
aggressive upscaling is taking place, this can significantly improve
texture detail in the final image.

* Combined Mip Bias Constants Into Vector

This change combines the mip bias shader constants into a single vector
rather than having two separate scalar values. This may improve
performance on older mobile devices and shouldn't hurt performance on
modern ones.

* 16-Bit FSR Cleanup & Fixes

This change removes the 16-bit option in the FSRCommon.hlsl in favor of
an automatic solution based on REAL_IS_HALF. Users of the FSR code no
longer need to explicitly opt-in to the 16-bit implementation. It will
now be enabled automatically whenever the target platform has support.

Some vendors have issues with 16-bit floating point casts to uint16
which are required for FSR's math approximations. This issue only
occurs on DX11 drivers, so the 16-bit implementation is currently
disabled whenever DX11 is in use until this is fixed.

This change also contains a small modification for the uint to float
conversion logic used inside FSRUtils.cs:SetRcasConstants. The updated
logic avoids an unnecessary memory allocation which was causing
failures in URP's graphics tests.

* Add FSR Upscaling Filter to URP

This change adds a new upscaling filter to URP that uses FidelityFX
Super Resolution. It is only available on platforms that support
shader target 4.5 or above. On unsupported platforms, it will behave
as if the user selected the "Auto" filter in the pipeline asset.

Unlike other filters, this filter will execute even when the renderer
is not configured to scale. HDRP has the same behavior. This allows the
FSR shaders to improve visual quality even when scaling is not active.

* Misc Backport Merge Fixup

This change removes some lines that were added by accident during the
merge process and also adds some logic to release temporary render
targets which were created during the backport process.

* Add Updated Validation Exception File

This file keeps getting automatically modified so it seems like it
should be updated as part of the backport.

* Fixed SetPerCameraShaderVariables

* * Moved `SetPerCameraShaderVariables` into percamera setup scope.
This call has to be called after SetupCameraProperties to reset the
values.

* * Moved SetCameraMatrices call inside the SetPerCameraShaderVariables

* * Updated CHANGELOG.md.

Co-authored-by: Felipe Lira <felipedrl@gmail.com>

* Update TextureGather Macros in GLCore.hlsl

This change updates the platform shader macros in GLCore.hlsl to enable
support for TextureGather operations on Unity's 4.5 shader target.

The changes in this PR are based off of GH #2346 which modified the
GLES3.hlsl file in order to fix cubemap array support. This PR basically
takes those same changes and applies them to GLCore.hlsl in addition.

This commit also adds some missing component specific macros for
texture gather (red, green, blue, alpha).

Co-authored-by: thomas-zeng <49886741+thomas-zeng@users.noreply.github.com>
Co-authored-by: Felipe Lira <felipedrl@gmail.com>