SG2 first pass asset implementation  (#7246)
This is the initial implementation which associates the ".sg2" extension with our own importer and editor. This also creates both a CLDS as well as a GTF asset that are associated with the file. This does not hook up any saving of changes that will need to be hooked up from GTF side. There is also still an issue that Sai is aware of with how the GTF window creates a new graph which he will be fixing in a follow up pr.

* asset work start

* Asset menu creation

* Fixup ShaderGraphImporterLegacy namechange for HDRP.

* Added a shim for a default registry to be shared by the asset importer and the stencil for the GTF editor-- this ensures a context node exists in CLDS whenever a graph is created-- this will need to be explored further next quarter when targets and context nodes are explored proper.

* undo-ing a name change to make merges easier later

* importers only work if class name matches file name. neat.

* Revert HDRP's rename for callsites

* Update MaterialPostProcessor.cs

* whitespace nit

* addressing Sai's comments

* merge conflict fix

* making GraphType public temporarily for compilation

Co-authored-by: Esmeralda Salamone <esmeralda.salamone@unity3d.com>
Co-authored-by: Esmeralda Salamone <esmelusina@gmail.com>