Update iOS URP tests
We are attempting to make the iOS builders for URP available as part of QV. As a pre-step we need to get them green and update them. This PR updates the reference images to the latest as well as disables some failing tests. The builder can be seen below that shows the iOS URP tests are now passing.

https://katana.ds.unity3d.com/projects/Unity/builders/proj0-PlaymodeURPTests%20-%20iOS%20%28iPhone%20SE%20-%20Metal%29%20-%20All%20parts/builds/38?unity_branch=graphics/fix-urp-tests