Graphics SRP - Add ref images for PS5 raytracing tests
- We are introducing a yamato pipeline for the PS5 raytracing tests in [ant-graphics-ci #74](https://github.cds.internal.unity3d.com/unity/ant-graphics-ci/pull/74)). 
This PR is adding the needed reference images, that don't exist yet in the repo (+meta files). ⚠ The ref images are taken from the win dx11 folder, which means that they may not match what we want on PS5. This PR intends to kick off the work on PS5 raytracing and a followup PR will update any mismatching image.
- Used this PR to also update the [package.lock.json](https://github.cds.internal.unity3d.com/unity/unity/pull/8884/files#diff-59904244e725d771e149c8d12737a7685659a8349e41020ce0f83dbd6ed3ac12) with the new unity paths (replacing the paths in use in the graphics repo).