Proper pivot on eye sample prefabs
**Starting point**
The eye model requires to have a pivot around the center of the pupils. 
**Problem**
When you want to animate the eye, the center of rotation is not properly placed, so it creates a wonky rotation.
**Solution**
Since we can't modify the model's pivot, the simple solution is to modify the prefab by adding a parent with a properly placed pivot (center of the eye). 

This PR does exactly that. 

**Before**
_Pivot Position_
![BeforeEyePivot](https://media.github.cds.internal.unity3d.com/user/1764/files/0d7ffb17-b354-469a-a136-419d51670fbc)
_Video_
https://media.github.cds.internal.unity3d.com/user/1764/files/6a5a3abc-6deb-40a7-affd-356999644ea1

**After**
_Pivot Position_
![ProperlyPlacedPivot](https://media.github.cds.internal.unity3d.com/user/1764/files/3caca63d-7d06-4dfc-8cdd-16f04a3ff6b4)
_Hierarchy_
![After-Eye](https://media.github.cds.internal.unity3d.com/user/1764/files/6e6cce75-667c-4820-b683-124986a27492)
_Video_
https://media.github.cds.internal.unity3d.com/user/1764/files/33f05b22-dd31-48da-a1d9-3aed999d9da6