Use packed depth-stencil format again to improve perf on some platforms
https://unity3d.atlassian.net/browse/TRND-2052

Performance regression was detected between 2020.3 and 2021.2 on a modified version of URP template.
After investigation, this is due to the depth buffer on Switch to be D32_S8 format instead of D24_S8 format. This regression happened several months ago since the C# API was fixed to allow creating depth render-target with 32bits depth.

In the test scene (see TRND-2052), using D32_S8 depth cost an extra 4ms (due to foliage overdraw).