[VFX] Fix crash due to incorrect Update order
When using GPU events, the emitter always needs to be updated before the receiver, which was not the case now. Currently the order depends on the order of creation.
Before it was not a problem, the error is only on the first frame and the initialization value, though invalid, didn't cause any visible issues.
However, we recently changed some initial value in this PR:
https://ono.unity3d.com/unity/unity/pull-request/135799/_/graphics/vfx/fix/1378335-gpuevent-initialization
This caused that, if the receiver system was created before the emitter, the first frame it could potentially access invalid memory and cause a crash.

The fix makes sure that the systems get ordered properly by layer.

The crash is 100% reproduced on DX12, using the repro steps in fogbugz