Bump graphics packages to 13.1.8
According to what was discussed with the graphics packages stakeholders, source control and RM here https://unity.slack.com/archives/GAULB1M5J/p1647601595159909?thread_ts=1647514831.342049&cid=GAULB1M5J, this PR is a manual attempt as bumping the graphics packages, since there is no automation yet.

The packages are bumped from 13.1.7 to 13.1.8.
A new section was added to the changelogs, they will need to be actually filled _after_  this PR lands, as we don't know yet what will make it to 13.1.8.