Added an attribute to support logarithmic sliders
POI: https://jira.unity3d.com/browse/RND-159

In the VFX Graph nodes UI is defined with attributes. A new coming feature needs a **logarithmic** slider to cover a specific range of values (power of 2). So this PR adds a generic and reusable way to define a logarithmic slider.

The purpose of a logarithmic slider is that the distance between each power of the base (10, 2 or even e) is constant.
In the capture below the slider is setup with a logarithmic scale on base 2 with a minimum to 1 and maximum to 2048.
We can see that the distance between 2^0 (1), 2^1 (2), 2^2 (4), 2^3 (8) ... 2^11 (2048) is constant.
![logarithmic-scale-base2](https://media.github.cds.internal.unity3d.com/user/4003/files/9c3075cd-6979-45f3-92ef-57325a9fd87e)