Graphics Packages tests - Bump gfx test framework to 7.8.23-exp.1
This PR is bumping the gfx-test-framework dependencies of the graphics packages test projects.
By doing so, we are fetching some updates that allow us to fix the reference images not found issue for xboxseries tests ([in sync with this PR](https://github.cds.internal.unity3d.com/unity/ant-graphics-ci/pull/155)).