Graphics/vfx/fix/1367447 improve inspector preview
https://fogbugz.unity3d.com/f/cases/1367447
Original [PR](https://github.com/Unity-Technologies/Graphics/pull/7168/)

When a VFX asset is selected the whole editor hits a frame rate drop.

After investigation it appeared that this slow down was mainly due to the VFX preview. See capture below:
![BHdjx9k9KO](https://user-images.githubusercontent.com/85612173/154319038-819ca463-7ed7-4af6-9201-07d73827dcd3.png )

To improve performances a new toggle button is available in the preview title bar to enable/disable animation.
By default it's disable so that when a VFX asset is selected there's no frame rate drop.
You can see in the capture below that the CPU usage is quite stable even after selecting the asset.
![QSWGoKlt8g](https://user-images.githubusercontent.com/85612173/154320459-b5c41cf4-a876-45e1-bb86-aa7d3461035d.png )