[RPW][HDRP]Multiple fixes on the Rendering Debugger
This PR aims to solve multiple issues with the Rendering Debugger and it's implementation on HDRP.

- [[HDRP] Entering Playmode with any Rendering Debugger mode on resets all debug view](https://fogbugz.unity3d.com/f/cases/1401599/)

Partially solving this issue, as the current limitations of the framework do not allow how the HDRP is using it. There is a splitted enum across multiple properties that when it is being set, multiple fields are reset or changed. For those to work properly when entering playmode, the enum must be splitted, and as talked with @julien this will be handled by HDRP team.

- [[HDRP] Undo action name quickly changes after Rendering Debugger UI is rebuilt](https://fogbugz.unity3d.com/f/cases/1412467/)
- [[HDRP] Some Rendering Debug actions are named as HDR: true in undo stack](https://fogbugz.unity3d.com/f/cases/1412465/)

The GetReset Method only destroys the IDebugData, but doesnt call the RegisterDebug UI from HDRP, as the new system is not reliying on the recreation of the UI to hide( not add) or show ( add) widgets, this is working out of the box correctly.
- [[HDRP] Reset button in Rendering Debugger doesn't update widget visibility](https://fogbugz.unity3d.com/f/cases/1411399/)

The main issue with this fogbugz was that all the implementation on the Rendering Debugger on HDRP is registering to the event `onValueChanged` as when it was implemented the callback `isHiddenCallback `did not exist.

- Drawers code for editor has been improved and shared.