Deprecate API related to Enlighten baking backend (v3)
I opened a new iteration since the previous was rotting and I needed to make a small change to URP tests. This PR was previously approved in principle, so I hope we can get this through fairly fast :)

This PR addresses https://jira.unity3d.com/browse/LIGHT-726 which is about marking API endpoints related to Enlighten baking as deprecated in order to signal that they will go away soon.

The PR does not change behavior - that comes later, instead, it just marks API as deprecated with the attribute set to cause warnings instead of errors.

Since the generated VS solutions treat deprecation warnings as errors by default, I've disabled the relevant warning in affected places. This seems to be a pretty common pattern throughout the codebase. Later, the diff of this PR could be a help to find places that should be changed when the backend is fully obsoleted.