Fixed a warning on first frame when a realtime probe is set to OnEnable
Fixed a warning on first frame when a realtime probe is set to OnEnable.
On first frame or when ambient probe is not ready, we "skip" rendering of the probes and repeat until ambient is ready.
With the current code, we do the rendering even when we skip but we would not set properly the render target which generated a warning.

https://unity-ci.cds.internal.unity3d.com/job/13168795/dependency-graph