URP RenderGraph foundation work (v2)
Jira: https://jira.unity3d.com/browse/URP-1370
RND Jira:  https://jira.unity3d.com/browse/RND-175

This PR prepares the ground for the URP RenderGraph work:

- Added an hidden option to toggle the RenderGraph render path (use the RENDER_GRAPH_ENABLED to enable the checkbox in the UI)
- All the passes, except the post processing ones, have been refactored and ported to RenderGraph. All the rendering code was moved into static functions so it can be called from both the current SRP path and the new RenderGraph path. As part of the refactoring, all the SetRenderTarget code has been moved out of the shared static function (since RenderGraph will take care of that), and is called directly in the SRP Execute().
- Added the infrastructure for the RenderGraph render path, this provides new base methods for the ScriptableRenderers to override
- UniversalRendererRenderGraph provides a base implementation for the RenderGraph path of the UniversalRenderer. NOTE: this file is currently containing all passes scheduled with no optimizations or culling, since it is the current basic frame used for porting passes. All this code is totally hidden from users, not possible to enable (the UI is hidden), and it is just used by the development team to keep working on rebuilding the full URP frame. So it doesn't need testing since it is work in progress and not final.

The main goals of this PR are: to provide the (hidden) render graph render path, to refactor all basic internal passes to be "RenderGraph friendly" without changing any of their functionality

NOTE TO QA: the RenderGraph path shouldn't be tested since it still doesn't provide the same functionality as the old path, since the work is not complete yet. It is very important to test the "old" non RenderGraph path to make sure that all the refactoring didn't change any of the functionality or the results of the existing passes.