[URP] Case 1405209 recorder capture fix
Fixes these issues:
https://fogbugz.unity3d.com/f/cases/1405209/
https://fogbugz.unity3d.com/f/cases/1405224/

Introducing the swap buffer system caused a recorder regression, since the recorder source texture wasn't set up properly at Setup time.
Because of the dynamic nature of the swap buffer, the capture pass source texture needs to be setup at Execute time so we are sure we are accessing the correct swap buffer.