[HDRP][Path Tracing] Replaced recursive continuation rays with loop in ray generation (v3)
Although user-wise there should not be too much visible change (if at all), other than slightly improved performance (in the 5-10% range after early tests), this PR removes the recursion from the path tracing, with no more continuation rays shot from closesthit().

This allow us to use an arbitrary number of bounces (still caped to 32 in the UI, but could be increased to any value now), without adding to the recursion level. This also makes aspects of the code cleaner and easier to read (no more do we have to deal with different miss() functions, for instance).

This will also hopefully open the door to other optimizations, like returning a BSDFData-like payload from closesthit() and do all shading computations in the raygen(), or even return a materialID and compute BSDFData or equivalent in the raygen() too.

Note that this PR existed on the public repository already, and was validated by all reviewers, but slightly too late to be merged before the repo change: https://github.com/Unity-Technologies/Graphics/pull/7090