Finding a view’s projection type during BRG culling (v2)
[RFC: Finding a view’s projection type during BRG culling](https://docs.google.com/document/d/1T0VlP0n7keHPVXmLHuXnJrZJ3-VnKCi1zwzHkhOSHQI/edit?usp=sharing)

The BatchRenderGroup (BRG) API offers a code path for culling that is different from the way most other things in Unity are culled.

![image](https://media.github.cds.internal.unity3d.com/user/2726/files/99d0d3fe-ea45-4673-9a32-1a71fff45bf8)

In this code path, culling is done in C# and hence is scriptable by the user. Without the BRG, culling is done in C++ and is not scriptable.

In the BRG API, we pass in a culling matrix for each view being culled. When doing culling in C#, the user will want to know what kind of projection that view has. This PR adds an API to do so.

The essence of this PR is adding this member to the context struct:

```
struct BatchCullingContext
{
    // Add this member variable
    BatchCullingProjectionType projectionType;
}
```

The rest of the code is to allow SRPs to provide this information to the engine. This is needed because we cannot infer projection type purely based on the light's type. E.g. even if spot lights usually have perspective projection, HDRP's _box-shaped_ spot lights are orthographic.