Add support for skinned motion vectors in HDRP for DOTS (v3)
This PR enables a path for motion vectors in HDRP for skinned entities. It changes the code that is generated from the Compute Deform node in Shader Graph.

Adds a path in MotionVector vertex shader when dots instancing and deformation defines are enabled. The previous positions are then overwritten by data from DOTS buffers.
The function is behind the HAVE_MESH_MODIFICATION define inside HDRP. This is enabled by the SG node generation when position, normal or tangent is modified. This is always true for our path as we rely on our Compute Deform node. The Compute Deform node also guarantees that vertex ID is defined.

Compute Deformation node now includes a file from entities graphics and adds a DOTS_DEFORMATION shader define. For now this is hidden behind a motion vector define because it changes the material property from uint to uint4. The file it includes contain the shader logic for applying the deformed vertex data.
Linear Blend Skinning node uses ByteAdressBuffer instead of Structured. This is fixing an existing issue in DOTS Deformations code for LBS.

Also removed old define to enable dots deformations which is no longer relevant.


Note: This is a simpler version of the code implemented as a fix for this bug https://fogbugz.unity3d.com/f/cases/1296266/