[HDRP] Transmission mask support - better scattering distance control
https://jira.unity3d.com/browse/RND-159

* Transmission Mask

When doing mixed materials, or a material atlas, it was previously not possible to mask transmission in some parts (actually you could put crazy high values in thickness texture but it's weird). For example for trees you only want transmission on leaves and not on the rest of the tree.

added input in all lit shaders + lit/stacklit/fabric shadergraph to control transmission intensity.
It's similar to subsurface mask, but it's only for transmission. Added graphic test for that

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/763e4e92-ce21-4622-987b-15288ff62eef)


* Scattering distance

Artist request - scattering distance was very hard to control because of broken picker, and it's actually not intuitive to put something HDR here

| before | after (auto migrated) |
| -- | -- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/98397e3a-b639-46cb-bdd7-dcb591de0e25) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/d26e3ce7-57b0-416f-99e5-8085bd18fc1f) |

* Added error box

When no profile is assigned, added an error box explaining why the material is purple
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/41fa00aa-bdeb-4e3d-989f-53865a5d5cef)