[VFX] Activation slots in blocks
This PR implements the activation ports in block to allow a better/more convenient way to enable/disable blocks with graph.
Activation expressions can be evaluated on both CPU and GPU and works for spawner blocks. Also for compile time constant, an optimized path is used.

![activationslot](https://user-images.githubusercontent.com/22473360/137947511-7dbeb592-6c4f-4534-bf6b-93afc5c16902.png         )

In addition to the core feature, this PR also:

- Fixes some useless recompilation triggered when dragging an edge out of an output slot connected to some compilable system
- Add bool <> uint, int, float casts
- Optimize the reduction of logical expressions

Fixed https://fogbugz.unity3d.com/f/cases/1428639/