[HDRP] Add scenes about stacking in transparency section of material samples
The detail of the work is in this [document](https://docs.google.com/document/d/1OT88f9wOhS5Wi2oclMxqItvic3EFQXFTr2BuE6DxuYQ/). 

This is a lighter version of [this now closed PR](https://github.cds.internal.unity3d.com/unity/unity/pull/8476). 

This PR specifically takes care of the stacking part. 

![A  Stacking - Raster - Transparency_20220315_1232691_Raster](https://media.github.cds.internal.unity3d.com/user/1764/files/69f66ecf-3b95-4048-9b09-0be893144753)
***
![B  Stacking - Ray Traced - Transparency_20220315_1232427_Ray Traced](https://media.github.cds.internal.unity3d.com/user/1764/files/7e5eea33-4af8-4b7a-aa20-1309f474a635)
***
![C  Stacking - Path Traced - Transparency_20220315_1232501_Path Traced](https://media.github.cds.internal.unity3d.com/user/1764/files/aca8823a-56df-42fa-8506-23242204d84e)

Also, a few improvement have been made about being able to see TMPPro text with the pathtracer (with a custom pass) so this has been retrofitted in the shadows scenes.
***
The naming of the scenes have changed a bit to accomodate two things: 
- The fact that Raster scene was not first when order alphabetically (now the order is Raster, Ray Tracing, Path Tracing) 
- The fact that Transparency was first in the name so when opened, we could only see "transparency" in the top bar.

_Before_
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/5c6a2fe3-e292-473a-b57e-72491bddbdaf)
_After_
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/b8f67265-a820-428b-b781-b8f8fff16402)

Maybe this will be done similarly for the material samples scenes (to start with the LIT) and also to put them in a folder called materials or something once all the PR are merged. 
***
Size: This adds only a few kb to the material samples (going from 43.37mb to 43.61mb)
 ***