Graphics/hdrp/eye cinematic caustic
This PR adds a new Eye material subtype which is an improvement on top of the existing Cinematic Eye -subtype. It adds an approximation of the caustic on the iris, induced by the cornea. In the image and in the video, left is the subtype introduced by this PR, middle is the cinematic eye, right is the eye subtype) . 

https://media.github.cds.internal.unity3d.com/user/3604/files/4672c92b-5a7a-4545-8df7-d45079c7c8a8

![image](https://media.github.cds.internal.unity3d.com/user/3604/files/0f1f48fc-bbf1-49ab-a908-c881d1a8f72c)

Jira: https://jira.unity3d.com/browse/HDRP-2327