[HDRP] Allow overwrite diffusion profile IOR on StackLit
https://jira.unity3d.com/browse/RND-159

Request by enemies demo. Currently, if a diffusion profile is assigned to a material, the fresnel is governed by the ior from the diffusion profile and overwrite whatever is set in the shadergraph. This PR removes this limitation by allowing to ignore diffusion profile IOR. This is disabled by default to keep compatibility.

Option in the stacklit graph inspector:
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/1e5697dc-19dd-4b8b-989f-36217d52a865)

When ticked (default), **Dielectric IOR** (or **Specular color** depending on **Base Color Parametrization**) are greyed out because they won't be used.
When unticked, these values become accessible

Currently when using specular base parametrization, there's an "Energy Conservation". But it's using the specular color so if we are getting the specular from the diffusion profile, this would be wrong. So i disabled the option when "use ior from profile" is ticked. This is a changed compared to trunk, but it's actually a fix

Mentionned it in stacklit doc, added entry in what's new.

Now:
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/09e7150a-bc93-4dc8-9bb3-409575a142fd)

Previously, the specular color didn't work
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/63f02d40-6e9b-4599-a198-ff4348e0fd42)
(but it still had an impact on the base color because of energy conservation stuff i wrote above)