Add option to use shadowmap as directional visibility for Marschner m…
This PR adds a possibility to choose between using a shadowmap or strand count probe as a directional visibility for multiscattering in hair SG node (the physical model). Before this, the hair would use the shadowmap visibility for punctual lights and probe for area lights.

Jira: https://jira.unity3d.com/browse/HDRP-2328