HDRP Cubemap Atlas
The primary purpose of this PR is to replace cube reflection probe cache array (Cubemap Array) with 2d texture atlas in octahedral projection. Which will allow us to be able to individually select cube probe resolution per probe. Which will improve performance and memory in different scenarios.
The second purpose of this PR is to combine planar and cube reflection probe atlas in a single 2d texture. Which will simplify code and improve memory usage. 

![2022-04-08 17_49_48-Administrator_ CubemapAtlas - _SampleScene - Windows, Mac, Linux - Unity 2022 2](https://media.github.cds.internal.unity3d.com/user/1206/files/f3ebb0d8-d8b7-4dca-85a5-65d159ab5e9b)
![2022-04-08 17_54_28-Administrator_ CubemapAtlas - _SampleScene - Windows, Mac, Linux - Unity 2022 2](https://media.github.cds.internal.unity3d.com/user/1206/files/03ef7ee4-d48b-4a97-81ec-44f59454cff2)
![2022-04-08 17_53_54-Administrator_ CubemapAtlas - _SampleScene - Windows, Mac, Linux - Unity 2022 2](https://media.github.cds.internal.unity3d.com/user/1206/files/1a765d9b-4b57-4277-b0bc-6ac5f9b3d41a)